chore: rename to element-helper
